+++ packages/web/app/views/patients/components/AccessorField.jsx
@@ -3,7 +3,7 @@ import { LAB_TEST_RESULT_TYPES } from '@tamanu/constants';
 import styled from 'styled-components';
 import { Field, NumberField, TextField } from '../../../components/Field';
 import { Colors } from '../../../constants';
-import { TranslatedOptionSelectField } from '../../../components/Translation/TranslatedOptionSelect';
+import { TranslatedOptionSelectField } from '../../../components/Translation/TranslatedOptions';
 
 const StyledField = styled(Field)`
   .Mui-disabled {
